More casts
Hg-commit: ac0361d5c48e5f0a51000bb43bc65dcdddf576ca

+++ src/dm/mesh/sieve/Mesh.hh
@@ -1656,7 +1656,7 @@ namespace ALE {
       if (this->depth() == 1) {
         ISieveTraversal<sieve_type>::orientedClosure(*this->getSieve(), p, rV);
       } else {
-        ISieveVisitor::PointRetriever<sieve_type,ISieveVisitor::RestrictVisitor<Section> > pV((int) pow(this->getSieve()->getMaxConeSize(), this->depth())+1, rV, true);
+        ISieveVisitor::PointRetriever<sieve_type,ISieveVisitor::RestrictVisitor<Section> > pV((int) pow((double) this->getSieve()->getMaxConeSize(), this->depth())+1, rV, true);
 
         ISieveTraversal<sieve_type>::orientedClosure(*this->getSieve(), p, pV);
       }
@@ -1670,7 +1670,7 @@ namespace ALE {
       if (this->depth() == 1) {
         ISieveTraversal<sieve_type>::orientedClosure(*this->getSieve(), p, uV);
       } else {
-        ISieveVisitor::PointRetriever<sieve_type,ISieveVisitor::UpdateVisitor<Section> > pV((int) pow(this->getSieve()->getMaxConeSize(), this->depth())+1, uV, true);
+        ISieveVisitor::PointRetriever<sieve_type,ISieveVisitor::UpdateVisitor<Section> > pV((int) pow((double) this->getSieve()->getMaxConeSize(), this->depth())+1, uV, true);
 
         ISieveTraversal<sieve_type>::orientedClosure(*this->getSieve(), p, pV);
       }
@@ -1696,7 +1696,7 @@ namespace ALE {
       if (this->depth() == 1) {
         ISieveTraversal<sieve_type>::orientedClosure(*this->getSieve(), p, uV);
       } else {
-        ISieveVisitor::PointRetriever<sieve_type,ISieveVisitor::UpdateAddVisitor<Section> > pV((int) pow(this->getSieve()->getMaxConeSize(), this->depth())+1, uV, true);
+        ISieveVisitor::PointRetriever<sieve_type,ISieveVisitor::UpdateAddVisitor<Section> > pV((int) pow((double) this->getSieve()->getMaxConeSize(), this->depth())+1, uV, true);
 
         ISieveTraversal<sieve_type>::orientedClosure(*this->getSieve(), p, pV);
       }
